detail 패키지수정

+++ src/main/java/com/green/winey_final/detail/model/SelSale.java
@@ -7,6 +7,6 @@ import lombok.Data;
 @AllArgsConstructor
 public class SelSale {
     private Long productId;
-    private Long sale;
-    private Long salePrice;
+    private Integer sale;
+    private Integer salePrice;
 }
